Build ginkgocadx on Debian 12
It builds successfully with new versions of deps (dcmtk, vtk, wxwidgets etc) on Debian 12, but only partial functionality is available, some of the functionality is broken.

Unfortunately, my plans do not include fixing everything that has broken due to new versions of dependencies, but at least it builds, maybe it will help someone.